requirements.txt
Django==4.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
mixer==7.2.2
requests==2.28.1
